Move ProtocolEngineExecuteError and adapt_protocol_source() to entrypoint_utils.
